renamed files and got everything working again
